git_proejct/requirements.txt
Flask
sentence-transformers
requests
beautifulsoup4
python-docx
PyMuPDF